@@ -1,33 +1,35 @@
 module github.com/weaveworks/progressive-delivery
 
-go 1.20
+go 1.22.0
+
+toolchain go1.22.5
 
 require (
 	github.com/bufbuild/buf v1.4.0
 	github.com/deepmap/oapi-codegen v1.12.4
-	github.com/fluxcd/flagger v1.30.0
-	github.com/fluxcd/helm-controller/api v0.30.0
-	github.com/fluxcd/kustomize-controller/api v0.34.0
-	github.com/fluxcd/source-controller/api v0.35.2
-	github.com/go-asset/generics v0.0.0-20220317100214-d5f632c68060
-	github.com/go-logr/logr v1.2.4
-	github.com/grpc-ecosystem/grpc-gateway/v2 v2.10.2
+	github.com/fluxcd/flagger v1.38.0
+	github.com/fluxcd/helm-controller/api v1.0.1
+	github.com/fluxcd/kustomize-controller/api v1.3.0
+	github.com/fluxcd/source-controller/api v1.3.0
+	github.com/go-asset/generics v0.0.0-20220617124532-a25edc3ff8d4
+	github.com/go-logr/logr v1.4.2
+	github.com/grpc-ecosystem/grpc-gateway/v2 v2.21.0
 	github.com/grpc-ecosystem/protoc-gen-grpc-gateway-ts v1.1.1
 	github.com/hashicorp/go-multierror v1.1.1
-	github.com/stretchr/testify v1.8.1
-	github.com/urfave/cli/v2 v2.8.1
-	github.com/weaveworks/weave-gitops v0.21.2
-	google.golang.org/genproto v0.0.0-20230410155749-daa745c078e1
-	google.golang.org/grpc v1.54.0
+	github.com/stretchr/testify v1.9.0
+	github.com/urfave/cli/v2 v2.27.3
+	github.com/weaveworks/weave-gitops v0.38.0
+	google.golang.org/genproto/googleapis/api v0.0.0-20240805194559-2c9e96a0b5d4
+	google.golang.org/grpc v1.65.0
 	google.golang.org/grpc/cmd/protoc-gen-go-grpc v1.1.0
-	google.golang.org/protobuf v1.30.0
+	google.golang.org/protobuf v1.34.2
 	gopkg.in/yaml.v3 v3.0.1
-	k8s.io/api v0.26.1
-	k8s.io/apiextensions-apiserver v0.26.1
-	k8s.io/apimachinery v0.26.1
-	k8s.io/client-go v0.26.1
-	k8s.io/utils v0.0.0-20230406110748-d93618cff8a2
-	sigs.k8s.io/controller-runtime v0.14.4
+	k8s.io/api v0.30.3
+	k8s.io/apiextensions-apiserver v0.30.3
+	k8s.io/apimachinery v0.30.3
+	k8s.io/client-go v0.30.3
+	k8s.io/utils v0.0.0-20240711033017-18e509b52bc8
+	sigs.k8s.io/controller-runtime v0.18.4
 	sigs.k8s.io/kustomize/kstatus v0.0.2
 )
 
