@@ -1,6 +1,6 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.27.1
+// 	protoc-gen-go v1.34.2
 // 	protoc        (unknown)
 // source: api/prog/prog.proto
 
@@ -807,7 +807,7 @@ func file_api_prog_prog_proto_rawDescGZIP() []byte {
 }
 
 var file_api_prog_prog_proto_msgTypes = make([]protoimpl.MessageInfo, 13)
-var file_api_prog_prog_proto_goTypes = []interface{}{
+var file_api_prog_prog_proto_goTypes = []any{
 	(*GetVersionRequest)(nil),           // 0: GetVersionRequest
 	(*GetVersionResponse)(nil),          // 1: GetVersionResponse
 	(*ListCanariesRequest)(nil),         // 2: ListCanariesRequest
@@ -866,7 +866,7 @@ func file_api_prog_prog_proto_init() {
 	}
 	file_api_prog_types_proto_init()
 	if !protoimpl.UnsafeEnabled {
-		file_api_prog_prog_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_api_prog_prog_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*GetVersionRequest); i {
 			case 0:
 				return &v.state
@@ -878,7 +878,7 @@ func file_api_prog_prog_proto_init() {
 				return nil
 			}
 		}
-		file_api_prog_prog_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+		file_api_prog_prog_proto_msgTypes[1].Exporter = func(v any, i int) any {
 			switch v := v.(*GetVersionResponse); i {
 			case 0:
 				return &v.state
@@ -890,7 +890,7 @@ func file_api_prog_prog_proto_init() {
 				return nil
 			}
 		}
-		file_api_prog_prog_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
+		file_api_prog_prog_proto_msgTypes[2].Exporter = func(v any, i int) any {
 			switch v := v.(*ListCanariesRequest); i {
 			case 0:
 				return &v.state
@@ -902,7 +902,7 @@ func file_api_prog_prog_proto_init() {
 				return nil
 			}
 		}
-		file_api_prog_prog_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
+		file_api_prog_prog_proto_msgTypes[3].Exporter = func(v any, i int) any {
 			switch v := v.(*ListCanariesResponse); i {
 			case 0:
 				return &v.state
@@ -914,7 +914,7 @@ func file_api_prog_prog_proto_init() {
 				return nil
 			}
 		}
-		file_api_prog_prog_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
+		file_api_prog_prog_proto_msgTypes[4].Exporter = func(v any, i int) any {
 			switch v := v.(*GetCanaryRequest); i {
 			case 0:
 				return &v.state
@@ -926,7 +926,7 @@ func file_api_prog_prog_proto_init() {
 				return nil
 			}
 		}
-		file_api_prog_prog_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
+		file_api_prog_prog_proto_msgTypes[5].Exporter = func(v any, i int) any {
 			switch v := v.(*GetCanaryResponse); i {
 			case 0:
 				return &v.state
@@ -938,7 +938,7 @@ func file_api_prog_prog_proto_init() {
 				return nil
 			}
 		}
-		file_api_prog_prog_proto_msgTypes[6].Exporter = func(v interface{}, i int) interface{} {
+		file_api_prog_prog_proto_msgTypes[6].Exporter = func(v any, i int) any {
 			switch v := v.(*IsFlaggerAvailableRequest); i {
 			case 0:
 				return &v.state
@@ -950,7 +950,7 @@ func file_api_prog_prog_proto_init() {
 				return nil
 			}
 		}
-		file_api_prog_prog_proto_msgTypes[7].Exporter = func(v interface{}, i int) interface{} {
+		file_api_prog_prog_proto_msgTypes[7].Exporter = func(v any, i int) any {
 			switch v := v.(*IsFlaggerAvailableResponse); i {
 			case 0:
 				return &v.state
@@ -962,7 +962,7 @@ func file_api_prog_prog_proto_init() {
 				return nil
 			}
 		}
-		file_api_prog_prog_proto_msgTypes[8].Exporter = func(v interface{}, i int) interface{} {
+		file_api_prog_prog_proto_msgTypes[8].Exporter = func(v any, i int) any {
 			switch v := v.(*ListMetricTemplatesRequest); i {
 			case 0:
 				return &v.state
@@ -974,7 +974,7 @@ func file_api_prog_prog_proto_init() {
 				return nil
 			}
 		}
-		file_api_prog_prog_proto_msgTypes[9].Exporter = func(v interface{}, i int) interface{} {
+		file_api_prog_prog_proto_msgTypes[9].Exporter = func(v any, i int) any {
 			switch v := v.(*ListMetricTemplatesResponse); i {
 			case 0:
 				return &v.state
@@ -986,7 +986,7 @@ func file_api_prog_prog_proto_init() {
 				return nil
 			}
 		}
-		file_api_prog_prog_proto_msgTypes[10].Exporter = func(v interface{}, i int) interface{} {
+		file_api_prog_prog_proto_msgTypes[10].Exporter = func(v any, i int) any {
 			switch v := v.(*ListCanaryObjectsRequest); i {
 			case 0:
 				return &v.state
@@ -998,7 +998,7 @@ func file_api_prog_prog_proto_init() {
 				return nil
 			}
 		}
-		file_api_prog_prog_proto_msgTypes[11].Exporter = func(v interface{}, i int) interface{} {
+		file_api_prog_prog_proto_msgTypes[11].Exporter = func(v any, i int) any {
 			switch v := v.(*ListCanaryObjectsResponse); i {
 			case 0:
 				return &v.state
