@@ -1,6 +1,6 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.27.1
+// 	protoc-gen-go v1.34.2
 // 	protoc        (unknown)
 // source: api/prog/types.proto
 
@@ -1589,7 +1589,7 @@ func file_api_prog_types_proto_rawDescGZIP() []byte {
 }
 
 var file_api_prog_types_proto_msgTypes = make([]protoimpl.MessageInfo, 19)
-var file_api_prog_types_proto_goTypes = []interface{}{
+var file_api_prog_types_proto_goTypes = []any{
 	(*Pagination)(nil),                 // 0: Pagination
 	(*ListError)(nil),                  // 1: ListError
 	(*Canary)(nil),                     // 2: Canary
@@ -1638,7 +1638,7 @@ func file_api_prog_types_proto_init() {
 		return
 	}
 	if !protoimpl.UnsafeEnabled {
-		file_api_prog_types_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_api_prog_types_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*Pagination); i {
 			case 0:
 				return &v.state
@@ -1650,7 +1650,7 @@ func file_api_prog_types_proto_init() {
 				return nil
 			}
 		}
-		file_api_prog_types_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+		file_api_prog_types_proto_msgTypes[1].Exporter = func(v any, i int) any {
 			switch v := v.(*ListError); i {
 			case 0:
 				return &v.state
@@ -1662,7 +1662,7 @@ func file_api_prog_types_proto_init() {
 				return nil
 			}
 		}
-		file_api_prog_types_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
+		file_api_prog_types_proto_msgTypes[2].Exporter = func(v any, i int) any {
 			switch v := v.(*Canary); i {
 			case 0:
 				return &v.state
@@ -1674,7 +1674,7 @@ func file_api_prog_types_proto_init() {
 				return nil
 			}
 		}
-		file_api_prog_types_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
+		file_api_prog_types_proto_msgTypes[3].Exporter = func(v any, i int) any {
 			switch v := v.(*CanaryTargetReference); i {
 			case 0:
 				return &v.state
@@ -1686,7 +1686,7 @@ func file_api_prog_types_proto_init() {
 				return nil
 			}
 		}
-		file_api_prog_types_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
+		file_api_prog_types_proto_msgTypes[4].Exporter = func(v any, i int) any {
 			switch v := v.(*CanaryStatus); i {
 			case 0:
 				return &v.state
@@ -1698,7 +1698,7 @@ func file_api_prog_types_proto_init() {
 				return nil
 			}
 		}
-		file_api_prog_types_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
+		file_api_prog_types_proto_msgTypes[5].Exporter = func(v any, i int) any {
 			switch v := v.(*CanaryCondition); i {
 			case 0:
 				return &v.state
@@ -1710,7 +1710,7 @@ func file_api_prog_types_proto_init() {
 				return nil
 			}
 		}
-		file_api_prog_types_proto_msgTypes[6].Exporter = func(v interface{}, i int) interface{} {
+		file_api_prog_types_proto_msgTypes[6].Exporter = func(v any, i int) any {
 			switch v := v.(*CanaryTargetDeployment); i {
 			case 0:
 				return &v.state
@@ -1722,7 +1722,7 @@ func file_api_prog_types_proto_init() {
 				return nil
 			}
 		}
-		file_api_prog_types_proto_msgTypes[7].Exporter = func(v interface{}, i int) interface{} {
+		file_api_prog_types_proto_msgTypes[7].Exporter = func(v any, i int) any {
 			switch v := v.(*FluxLabels); i {
 			case 0:
 				return &v.state
@@ -1734,7 +1734,7 @@ func file_api_prog_types_proto_init() {
 				return nil
 			}
 		}
-		file_api_prog_types_proto_msgTypes[8].Exporter = func(v interface{}, i int) interface{} {
+		file_api_prog_types_proto_msgTypes[8].Exporter = func(v any, i int) any {
 			switch v := v.(*Automation); i {
 			case 0:
 				return &v.state
@@ -1746,7 +1746,7 @@ func file_api_prog_types_proto_init() {
 				return nil
 			}
 		}
-		file_api_prog_types_proto_msgTypes[9].Exporter = func(v interface{}, i int) interface{} {
+		file_api_prog_types_proto_msgTypes[9].Exporter = func(v any, i int) any {
 			switch v := v.(*CanaryAnalysis); i {
 			case 0:
 				return &v.state
@@ -1758,7 +1758,7 @@ func file_api_prog_types_proto_init() {
 				return nil
 			}
 		}
-		file_api_prog_types_proto_msgTypes[10].Exporter = func(v interface{}, i int) interface{} {
+		file_api_prog_types_proto_msgTypes[10].Exporter = func(v any, i int) any {
 			switch v := v.(*CanaryMetric); i {
 			case 0:
 				return &v.state
@@ -1770,7 +1770,7 @@ func file_api_prog_types_proto_init() {
 				return nil
 			}
 		}
-		file_api_prog_types_proto_msgTypes[11].Exporter = func(v interface{}, i int) interface{} {
+		file_api_prog_types_proto_msgTypes[11].Exporter = func(v any, i int) any {
 			switch v := v.(*CanaryMetricThresholdRange); i {
 			case 0:
 				return &v.state
@@ -1782,7 +1782,7 @@ func file_api_prog_types_proto_init() {
 				return nil
 			}
 		}
-		file_api_prog_types_proto_msgTypes[12].Exporter = func(v interface{}, i int) interface{} {
+		file_api_prog_types_proto_msgTypes[12].Exporter = func(v any, i int) any {
 			switch v := v.(*CanaryMetricTemplate); i {
 			case 0:
 				return &v.state
@@ -1794,7 +1794,7 @@ func file_api_prog_types_proto_init() {
 				return nil
 			}
 		}
-		file_api_prog_types_proto_msgTypes[13].Exporter = func(v interface{}, i int) interface{} {
+		file_api_prog_types_proto_msgTypes[13].Exporter = func(v any, i int) any {
 			switch v := v.(*MetricProvider); i {
 			case 0:
 				return &v.state
@@ -1806,7 +1806,7 @@ func file_api_prog_types_proto_init() {
 				return nil
 			}
 		}
-		file_api_prog_types_proto_msgTypes[14].Exporter = func(v interface{}, i int) interface{} {
+		file_api_prog_types_proto_msgTypes[14].Exporter = func(v any, i int) any {
 			switch v := v.(*GroupVersionKind); i {
 			case 0:
 				return &v.state
@@ -1818,7 +1818,7 @@ func file_api_prog_types_proto_init() {
 				return nil
 			}
 		}
-		file_api_prog_types_proto_msgTypes[15].Exporter = func(v interface{}, i int) interface{} {
+		file_api_prog_types_proto_msgTypes[15].Exporter = func(v any, i int) any {
 			switch v := v.(*UnstructuredObject); i {
 			case 0:
 				return &v.state
@@ -1830,7 +1830,7 @@ func file_api_prog_types_proto_init() {
 				return nil
 			}
 		}
-		file_api_prog_types_proto_msgTypes[16].Exporter = func(v interface{}, i int) interface{} {
+		file_api_prog_types_proto_msgTypes[16].Exporter = func(v any, i int) any {
 			switch v := v.(*Condition); i {
 			case 0:
 				return &v.state
